src/requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
biothings-client==0.2.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
fastcluster==1.1.25
h5py==2.9.0
idna==2.8
importlib-metadata==0.19
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
mygene==3.1.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nose==1.3.7
notebook==6.0.0
numexpr==2.7.0
numpy==1.17.0
packaging==19.1
palettable==3.2.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
pyerf==1.0.1
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.3
requests==2.22.0
rpy2==3.0.5
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
statsmodels==0.10.1
tables==3.5.2
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.33.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.1.8
zipp==0.5.2
